feat(channels): standard channel order and full channel numbers
Browse order uses the SDK's standard TVHeadend channel comparator, so
channels without a number (TVHeadend reports 0) follow the numbered ones
by name instead of leading the list in id order. Channel number 0 is no
longer shown or matched, positional numbering applies when no channel has
a visible number, numbers stay Long end to end, and number entry accepts
as many digits as the largest reachable channel number.

+++ app/src/androidTest/java/at/bernhardberger/tvhplayer/ui/components/ChannelRowTest.kt
@@ -84,7 +84,7 @@ class ChannelRowTest {
                             modifier = Modifier
                                 .width(400.dp)
                                 .testTag("status-row-$index"),
-                            number = index + 1,
+                            number = index + 1L,
                             name = if (index == 2) {
                                 "Short title"
                             } else {
@@ -144,7 +144,7 @@ class ChannelRowTest {
                         val number = index + 1
                         ChannelRow(
                             modifier = Modifier.testTag("row-$number"),
-                            number = number,
+                            number = number.toLong(),
                             name = "Channel $number",
                             programTitle = "Programme $number",
                             progress = 0.5f,
@@ -185,7 +185,7 @@ class ChannelRowTest {
                         SideEffect { compositionCounts[number - 1]++ }
                         ChannelRow(
                             modifier = Modifier.testTag("measured-row-$number"),
-                            number = number,
+                            number = number.toLong(),
                             name = "Channel $number",
                             programTitle = "Programme $number",
                             progress = 0.5f,
